add normalize_interface in module_utils and fix nxos_l3_interface module
